raft node and main split into two files
