+++ cmd/main.go
@@ -6,15 +6,15 @@ import (
 )
 
 func main(){
-    /*cfg := raft.RawConfig{
+    cfg := raft.RawConfig{
         BindAddress: "192.168.0.1",
         JoinAddress: "192.168.0.1",
         RaftPort: 1000,
 		HTTPPort: 8080,
 		DataDir: "./raft",
 		Bootstrap: false,
-    }*/
-	cfg := raft.ReadRawConfig()
+    }
+	//cfg := raft.ReadRawConfig()
     result, err := cfg.ValidateConfig()
     fmt.Println(result)
     fmt.Println(err)
